chore: Extract PR to HOC
Reviewer does not want to merge the feature into `gatsby-image`. Extracted to separate a file as a Higher-Order Component.

Reverts changes to `index.js`. Loses key prop improvement with `activeVariant`. May not integrate as smoothly as-is due to changes in hydration fix rendering, may need to keep `uniqueKey` until hydrated.

To use, fluid or fixed prop is **expected** to be provided. Import the HOC version `gatsby-image/art-direction` and use as you would `gatsby-image`.
